refactor lib2 cookie class
